backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
google-generativeai==0.3.2
rapidfuzz==3.6.1
httpx==0.28.1
python-multipart==0.0.6
python-telegram-bot==22.2
requests==2.31.0
